Step 4 fix recursive functions
This required a lot more learning of borrowing etc. Now we can pass
around an immutable reference to a struct which has a RefCell allowing
internal mutation. This reduces the need for clone significantly,
maybe even altogether, although not yet.

Now immediate invocation of anonymous functions is broken, but it
feels like getting that to work in this new approach is tenable at
least.