Step 4 also fix immediate invocation of anonymous functions
The closure needs a strong reference to the env otherwise the env can
be dropped before the function is called. The trick was using
Rc::clone to increment the reference count to the env.

The final step would be to remove the cloning that happens with the
MalEnv methods, but that can come later.