Step 4 remove env clones
Now we rely on Rc for tracking references to the env rather than
cloning it every time a new env is needed. This means the memory will
better reflect the usage of the values in the program. May now also
consider applying Rc to the use of MalVal and clone, but that can wait.