Step 8 implementation of macros
With a better understanding of ownership the code, if you'll forgive
the pun, moves back to using Rc for managing the ast in
EVAL/eval_ast. This means some more cloning but also the code feels
easier to follow and also that macros could _actually_ be implemented,
this was the point at which Cell stopped being a viable solution for
this use case. Also makes TCO possible for user functions.