Simplify Rc env usage in EVAL
Using a RefCell to mutable hold onto a reference to a smarter pointer
proved too convoluted such that TCO for user functions didn't
work. When trying to replace the env with another it would panic with
'already borrowed: BorrowMutError' (as it was indeed borrowed
earlier). Going with mut and a plain ol' Rc solves all that and now
TCO for user functions works!