Step 4 Implement string functions and string escaping
Possibly one of the fiddliest parts of mal, and certainly where I've
spent a significant amount of time trying to get right. Now it passes
all but one of the step 4 tests:

    "\\n"
    ;=>"\\n"

Currently it produces "\\\n" as the string is read as "backslash
followed by a newline" rather than "backslash followed by an n". This
could be fixed with a scanning replacement (e.g s///g) but I can't see
an easy way to do that (by reading the re & string docs at
least). Being an edge case I'm not too bothered.

Also it doesn't handle unbalanced quotes etc but also not bothered at
this point. Maybe I'll clean this up at the end.