Correctly handle string escaping
While the strings looked sane when printed out they were actually a
jumble of lists of characters and "strings" under the hood e.g

    user> (def! p (pr-str {:abc "val1" :def "val2"}))
    "{:abc \"val1\" :def \"val2\"}"
    user> p
    "{:abc \"val1\" :def \"val2\"}"
    user> (dbg "{:abc \"val1\" :def \"val2\"}")
    {mal_str,"{:abc \\\"val1\\\" :def \\\"val2\\\"}"}
    nil
    user> (dbg p)
    {mal_str,[123,58,97,98,99,32,34,"val1",34,32,58,100,101,102,32,34,"val2",34,125]}

The bugs were twofold — one, the reader wasn't correctly escaping the
escapes for the regex replace and two, the printer was also not
escaping correctly but it was also producing an internal jumble which
would, not unreasonably, make equivalence checks fail. My lack of
understanding of how strings actually work in Erlang to begin (and at
least less so now) was the real source of the bugs!