Step 3 - remember that Map exists
There I was writing TypeScript like it was 2009. One caveat with this
is that access won't work except outside of def! values. This is due
to "equivalent" objects not being considered equivalent e.g
```typescript
{ type: 'keyword': value: 'foo' } === { type: 'keyword': value: 'foo'} //= false
```
At the point map access is needed I'll likely apply Symbol in some
kind of clever way.

Originally Map was also applied to Env but, without another layer of
indirection, it didn't provide any value so the code was rewritten back
to using a plain object again.